% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createAbalone}
\alias{createAbalone}
\title{Abalone Age prediction}
\usage{
createAbalone(file = "abalone.rds", write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write data file to}

\item{write}{logical; should the dataset be written to disk for later use? (default: TRUE)}

\item{read}{logical; should we try to read the dataset from the specified location first? (default: TRUE)}
}
\value{
The dataset as a \code{\link[data.table]{data.table}}
}
\description{
Predict the age of abalone from physical measurements
}
\details{
\emph{Task:} Regression: formula(Rings~.)
Often the  final 1044 examples as used test, the first 3133 for training.

\emph{From UCI:} "Data can be treated as a classification task, for instance Data set treated as a 3-category classification problem (grouping ring classes 1-8, 9 and 10, and 11 on)"
}
\references{
Warwick J Nash, Tracy L Sellers, Simon R Talbot, Andrew J Cawthorn and Wes B Ford (1994) "The Population Biology of Abalone (_Haliotis_ species) in Tasmania. I. Blacklip Abalone (_H. rubra_) from the North Coast and Islands of Bass Strait", Sea Fisheries Division, Technical Report No. 48 (ISSN 1034-3288)
}
\seealso{
\url{https://archive.ics.uci.edu/ml/datasets/Abalone}
}

