% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createAccelerometerActivity}
\alias{createAccelerometerActivity}
\title{Create Accelerometer Activity Recognition dataset}
\usage{
createAccelerometerActivity(file = "accelerometeractivity.rds",
  write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write data file to}

\item{write}{logical; should the dataset be written to disk for later use? (default: TRUE)}

\item{read}{logical; should we try to read the dataset from the specified location first? (default: TRUE)}
}
\value{
The dataset as a \code{\link[data.table]{data.table}}
}
\description{
From UCI: Uncalibrated Accelerometer Data are collected from 15 participantes performing 7 activities. The dataset provides challenges for identification and authentication of people using motion patterns. The dataset collects data from a wearable accelerometer mounted on the chest. Sampling frequency of the accelerometer: 52 Hz. Accelerometer Data are Uncalibrated. Number of Participants: 15. Number of Activities: 7
}
\details{
Task:  formula(Activity ~ . -Sequence - Participant)
}
\references{
Casale, P. Pujol, O. and Radeva, P. 'Personalization and user verification in wearable systems using biometric walking patterns' Personal and Ubiquitous Computing, 16(5), 563-580, 2012
}
\seealso{
\url{https://archive.ics.uci.edu/ml/datasets/Activity+Recognition+from+Single+Chest-Mounted+Accelerometer}
}

