% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAmazonSentiment.R
\name{createAmazonSentiment}
\alias{createAmazonSentiment}
\title{Create Amazon Sentiment dataset}
\usage{
createAmazonSentiment(file = getfilepath("amazonsentiment.rds"),
  write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write data file to}

\item{write}{logical; should the dataset be written to disk for later use? (default: TRUE)}

\item{read}{logical; should we try to read the dataset from the specified location first? (default: TRUE)}
}
\value{
List containing:
\itemize{
 \item{"X"}{ \code{dgCMatrix}; sparse matrix with count of unigram and bigram features}
 \item{"y"}{ factor; labels}
 \item{"domains"}{ factor; domain/category for each review}
}
}
\description{
The goal in this dataset is to predict binary (positive/negative) sentiment from product reviews taken from Amazon.com in four different categories: books, dvds, electronics and kitchen.
}
\details{
For 27677 reviews, 1110349 unigram and bigram features are given in a sparse matrix X, along with a label vector y and a vector indicating the domain. Note the data is NOT a \code{\link{data.table}}, but sparse matrix generated by \code{\link[Matrix]{sparseMatrix}}.

\emph{Task:} Classification: Use X to predict y, possibly in a domain adaptation setting.
}
\seealso{
\code{\link{createAmazonSentimentStars}}, \url{http://www.cs.jhu.edu/~mdredze/datasets/sentiment/}
}

