% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAmazonSentimentStars.R
\name{createAmazonSentimentStars}
\alias{createAmazonSentimentStars}
\title{Create Amazon Sentiment star ratings dataset}
\usage{
createAmazonSentimentStars(file = getfilepath("amazonsentimentstars.rds"),
  write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write data file to}

\item{write}{logical; should the dataset be written to disk for later use? (default: TRUE)}

\item{read}{logical; should we try to read the dataset from the specified location first? (default: TRUE)}
}
\value{
List containing:
\itemize{
 \item{"X"}{ \code{dgCMatrix}; sparse matrix with count of unigram and bigram features}
 \item{"y"}{ numeric; star rating}
 \item{"domains"}{ factor; domain/category for each review}
}
}
\description{
The goal in this dataset is to predict the number of stars given in a product review taken from Amazon.com in four different categories: books, dvds, electronics and kitchen.
}
\details{
For 21669 reviews, 1009925 unigram and bigram features are given in a sparse matrix X, along with a vector y containing the star rating and a vector indicating the domain. Note the data is NOT a \code{\link{data.table}}, but sparse matrix generated by \code{\link[Matrix]{sparseMatrix}}.

\emph{Task:} Classification: Use X to predict y, possibly in a domain adaptation setting.
}
\seealso{
\code{\link{createAmazonSentiment}}, \url{http://www.cs.jhu.edu/~mdredze/datasets/sentiment/}
}

