% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createAnnealing}
\alias{createAnnealing}
\title{Create Steel Annealing dataset}
\usage{
createAnnealing(file = "annealing.rds", write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write data file to}

\item{write}{logical; should the dataset be written to disk for later use? (default: TRUE)}

\item{read}{logical; should we try to read the dataset from the specified location first? (default: TRUE)}
}
\value{
The dataset as a \code{\link[data.table]{data.table}}
}
\description{
\emph{Task:} First 798 training, last 100 testing. Classification formula(Class~.)
}
\details{
\emph{From UCI:} "Explanation: I suspect this was left by Ross Quinlan in 1987 at the 4th Machine Learning Workshop.  I'd have to check with Jeff Schlimmer to double check this."
}
\examples{
randomForest(Class~.,na.roughfix(data[,which(sapply(1:39,function(x) {!all(is.na(data[,x,with=FALSE]))})),with=FALSE]))
}
\seealso{
\url{https://archive.ics.uci.edu/ml/datasets/Annealing}
}

