% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBalanceScale.R
\name{createBalanceScale}
\alias{createBalanceScale}
\title{Balance Scale dataset}
\usage{
createBalanceScale(file = getfilepath("balancescale.rds"), write = TRUE,
  read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write data file to}

\item{write}{logical; should the dataset be written to disk for later use? (default: TRUE)}

\item{read}{logical; should we try to read the dataset from the specified location first? (default: TRUE)}
}
\value{
The dataset as a \code{\link[data.table]{data.table}}
}
\description{
\emph{Task:} Multiclass: formula(Class ~ . )
}
\details{
\emph{From UCI:} "This data set was generated to model psychologicalexperimental results.  Each example is classified as having thebalance scale tip to the right, tip to the left, or bebalanced.  The attributes are the left weight, the leftdistance, the right weight, and the right distance.  Thecorrect way to find the class is the greater of (left-distance * left-weight) and (right-distance * right-weight).  If they are equal, it is balanced."
}
\references{
Siegler, R. S. (1976).  Three Aspects of CognitiveDevelopment.  Cognitive Psychology, 8, 481-520.
}
\seealso{
\url{https://archive.ics.uci.edu/ml/datasets/Balance+Scale}
}

