% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBanknote.R
\name{createBanknote}
\alias{createBanknote}
\title{Create Banknote Authentication dataset}
\usage{
createBanknote(file = getfilepath("banknote.rds"), write = TRUE,
  read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write data file to}

\item{write}{logical; should the dataset be written to disk for later use? (default: TRUE)}

\item{read}{logical; should we try to read the dataset from the specified location first? (default: TRUE)}
}
\value{
The dataset as a \code{\link[data.table]{data.table}}
}
\description{
From UCI: Data were extracted from images that were taken from genuine and forged banknote-like specimens. For digitization, an industrial camera usually used for print inspection was used. The final images have 400x 400 pixels. Due to the object lens and distance to the investigated object gray-scale pictures with a resolution of about 660 dpi were gained. Wavelet Transform tool were used to extract features from images.
}
\details{
1. variance of Wavelet Transformed image (continuous) 
2. skewness of Wavelet Transformed image (continuous) 
3. curtosis of Wavelet Transformed image (continuous) 
4. entropy of image (continuous) 
5. class (integer)

Task:  formula(Class ~ .)
}
\seealso{
\url{https://archive.ics.uci.edu/ml/datasets/banknote+authentication}
}

