% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createClimateSimulationCrashes}
\alias{createClimateSimulationCrashes}
\title{Create Climate Model Failure dataset}
\usage{

  createClimateSimulationCrashes(file = getfilepath("climatesimulationcrashes.rds"),
  write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write data file to}

\item{write}{logical; should the dataset be written to disk for later use? (default: TRUE)}

\item{read}{logical; should we try to read the dataset from the specified location first? (default: TRUE)}
}
\value{
The dataset as a \code{\link[data.table]{data.table}}
}
\description{
From UCI: "This dataset contains records of simulation crashes encountered during climate model uncertainty quantification (UQ) ensembles.Ensemble members were constructed using a Latin hypercube method in LLNL's UQ Pipeline software system to sample the uncertainties of 18 model parameters within the Parallel Ocean Program (POP2) component of the Community Climate System Model (CCSM4). Three separate Latin hypercube ensembles were conducted, each containing 180 ensemble members. 46 out of the 540 simulations failed for numerical reasons at combinations of parameter values. The goal is to use classification to predict simulation outcomes (fail or succeed) from input parameter values, and to use sensitivity analysis and feature selection to determine the causes of simulation crashes. Further details about the data and methods are given in the publication 'Failure Analysis of Parameter-Induced Simulation Crashes in Climate Models,' Geoscientific Model Development"
}
\details{
Task:  formula(outcome ~ . -Study -Run)
}
\seealso{
\url{https://archive.ics.uci.edu/ml/datasets/Climate+Model+Simulation+Crashes}
}

