% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createCOIL2}
\alias{createCOIL2}
\title{Create COIL2 dataset from the SSL benchmarks}
\usage{
createCOIL2(file = "COIL2.rds", write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write RData file to}

\item{write}{logical; should the dataset be written to disk? (default: FALSE)}
}
\description{
Task:
 formula(Class ~ .)
}
\details{
Cite: Chapelle
}

