% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createContraceptiveMethod}
\alias{createContraceptiveMethod}
\title{Create Contraceptive Method Choice dataset}
\usage{
createContraceptiveMethod(file = getfilepath("contraceptivemethod.rds"),
  write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write data file to}

\item{write}{logical; should the dataset be written to disk for later use? (default: TRUE)}

\item{read}{logical; should we try to read the dataset from the specified location first? (default: TRUE)}
}
\value{
The dataset as a \code{\link[data.table]{data.table}}
}
\description{
From UCI: This dataset is a subset of the 1987 National Indonesia Contraceptive Prevalence Survey. The samples are married women who were either not pregnant or do not know if they were at the time of interview. The problem is to predict the current contraceptive method choice (no use, long-term methods, or short-term methods) of a woman based on her demographic and socio-economic characteristics.
}
\details{
1. Wife's age (numerical)
2. Wife's education (categorical) 1=low, 2, 3, 4=high
3. Husband's education (categorical) 1=low, 2, 3, 4=high
4. Number of children ever born (numerical)
5. Wife's religion (binary) 0=Non-Islam, 1=Islam
6. Wife's now working? (binary) 0=Yes, 1=No
7. Husband's occupation (categorical) 1, 2, 3, 4
8. Standard-of-living index (categorical) 1=low, 2, 3, 4=high
9. Media exposure (binary) 0=Good, 1=Not good
10. Contraceptive method used (class attribute) 1=No-use, 2=Long-term, 3=Short-term

Task: two class: formula(factor(Contraceptive=="No-use") ~ .)
multiclass: formula(Contraceptive ~ .)
}
\references{
Lim, T.-S., Loh, W.-Y. & Shih, Y.-S. (1999). A Comparison of Prediction Accuracy, Complexity, and Training Time of Thirty-three Old and New Classification Algorithms. Machine Learning
}
\seealso{
\url{https://archive.ics.uci.edu/ml/datasets/Contraceptive+Method+Choice}
}

