% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createCovertype}
\alias{createCovertype}
\title{Create Covertype dataset}
\usage{
createCovertype(file = "covertype.rds", write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write data file to}

\item{write}{logical; should the dataset be written to disk for later use? (default: TRUE)}

\item{read}{logical; should we try to read the dataset from the specified location first? (default: TRUE)}
}
\value{
The dataset as a \code{\link[data.table]{data.table}}
}
\description{
From UCI: Predicting forest cover type from cartographic variables only (no remotely sensed data).  The actual forest cover type for a given observation (30 x 30 meter cell) was determined from US Forest Service (USFS) Region 2 Resource Information System  (RIS) data.  Independent variables were derived from data originally obtained from US Geological Survey (USGS) and USFS data.  Data is in raw form (not scaled)
 Task:
 formula(Cover_Type ~ .)
}
\references{
Blackard, Jock A. and Denis J. Dean.  2000.  "Comparative Accuracies of Artificial Neural Networks and Discriminant Analysis in Predicting Forest Cover Types from Cartographic Variables."  Computers and Electronics in Agriculture 24(3):131-151.
}
\seealso{
\url{https://archive.ics.uci.edu/ml/datasets/Covertype}
}

