% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{createdatasets}
\alias{createdatasets}
\alias{createdatasets-package}
\title{Create benchmark datasets}
\description{
This package allows one to download (if neccesary) and preprocess benchmark datasets from the UCI machine learning repository and other sources for the use in evaluating classification and regression methods.These functions can be helpful in setting up simulation studies that can be easily reproduced by others.
}

