% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createDiabetes}
\alias{createDiabetes}
\title{Pima Indians Diabetes dataset}
\usage{
createDiabetes(file = "diabetes.rds", write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write data file to}

\item{write}{logical; should the dataset be written to disk for later use? (default: TRUE)}

\item{read}{logical; should we try to read the dataset from the specified location first? (default: TRUE)}
}
\value{
The dataset as a \code{\link[data.table]{data.table}}
}
\description{
Diagnosis of diabetes in Pima Indian women (at least 21 years old)
}
\details{
\emph{Task}: formula(Diabetes~.)

\emph{Preprocessing}: Identify missing values and encode them as NA

\emph{From UCI description}: "Several constraints were placed on the selection of these instances from a larger database. In particular, all patients here are females at least 21 years old of Pima Indian heritage. ADAP is an adaptive learning routine that generates and executes digital analogs of perceptron-like devices. It is a unique algorithm; see the paper for details."
}
\references{
Smith, J.W., Everhart, J.E., Dickson, W.C., Knowler, W.C., & Johannes, R.S. (1988). Using the ADAP learning algorithm to forecast the onset of diabetes mellitus. In Proceedings of the Symposium on Computer Applications and Medical Care (pp. 261--265). IEEE Computer Society Press.
}
\seealso{
\url{https://archive.ics.uci.edu/ml/datasets/Pima+Indians+Diabetes}
}

