% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createDigit1}
\alias{createDigit1}
\title{Create Digit1 dataset from the SSL benchmarks}
\usage{
createDigit1(file = "digit1.rds", write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write RData file to}

\item{write}{logical; should the dataset be written to disk? (default: FALSE)}
}
\description{
Task:
 formula(Class ~ .)
}
\details{
Cite: Chapelle
}

