% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createEEGEye}
\alias{createEEGEye}
\title{Create EEG Eye State dataset}
\usage{
createEEGEye(file = "eegeye.rds", write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write data file to}

\item{write}{logical; should the dataset be written to disk for later use? (default: TRUE)}

\item{read}{logical; should we try to read the dataset from the specified location first? (default: TRUE)}
}
\value{
The dataset as a \code{\link[data.table]{data.table}}
}
\description{
From UCI: All data is from one continuous EEG measurement with the Emotiv EEG Neuroheadset. The duration of the measurement was 117 seconds. The eye state was detected via a camera during the EEG measurement and added later manually to the file after analysing the video frames. '1' indicates the eye-closed and '0' the eye-open state. All values are in chronological order with the first measured value at the top of the data.
}
\details{
Task: formula(eyeDetection ~ .)
}
\seealso{
\url{https://archive.ics.uci.edu/ml/datasets/EEG+Eye+State}
}

