% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFertility.R
\name{createFertility}
\alias{createFertility}
\title{Create Fertility Diagnosis dataset}
\usage{
createFertility(file = getfilepath("fertility.rds"), write = TRUE,
  read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write data file to}

\item{write}{logical; should the dataset be written to disk for later use? (default: TRUE)}

\item{read}{logical; should we try to read the dataset from the specified location first? (default: TRUE)}
}
\value{
The dataset as a \code{\link[data.table]{data.table}}
}
\description{
From UCI: Data were extracted from images that were taken from genuine and forged banknote-like specimens. For digitization, an industrial camera usually used for print inspection was used. The final images have 400x 400 pixels. Due to the object lens and distance to the investigated object gray-scale pictures with a resolution of about 660 dpi were gained. Wavelet Transform tool were used to extract features from images.
}
\details{
Season in which the analysis was performed. 1) winter, 2) spring, 3) Summer, 4) fall. (-1, -0.33, 0.33, 1) 
Age at the time of analysis. 18-36 (0, 1) 
Childish diseases (ie , chicken pox, measles, mumps, polio)  1) yes, 2) no. (0, 1) 
Accident or serious trauma 1) yes, 2) no. (0, 1) 
Surgical intervention 1) yes, 2) no. (0, 1) 
High fevers in the last year 1) less than three months ago, 2) more than three months ago, 3) no. (-1, 0, 1) 
Frequency of alcohol consumption 1) several times a day, 2) every day, 3) several times a week, 4) once a week, 5) hardly ever or never (0, 1) 
Smoking habit 1) never, 2) occasional 3) daily. (-1, 0, 1) 
Number of hours spent sitting per day ene-16	(0, 1) 
Output: Diagnosis	normal (N), altered (O)

Task:  formula(Diagnosis ~ .)
}
\references{
David Gil, Jose Luis Girela, Joaquin De Juan, M. Jose Gomez-Torres, and Magnus Johnsson. Predicting seminal quality with artificial intelligence methods. Expert Systems with Applications, 39(16), 2012
}
\seealso{
\url{https://archive.ics.uci.edu/ml/datasets/Fertility}
}

