% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createG241C}
\alias{createG241C}
\title{Create G241C dataset from the SSL benchmarks}
\usage{
createG241C(file = "g241c.rds", write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write RData file to}

\item{write}{logical; should the dataset be written to disk? (default: FALSE)}
}
\description{
Task:
 formula(Class ~ .)
}

