% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createHaberman}
\alias{createHaberman}
\title{Create Haberman breast cancer survival dataset}
\usage{
createHaberman(file = "haberman.rds", write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write data file to}

\item{write}{logical; should the dataset be written to disk for later use? (default: TRUE)}

\item{read}{logical; should we try to read the dataset from the specified location first? (default: TRUE)}
}
\value{
The dataset as a \code{\link[data.table]{data.table}}
}
\description{
Survival of patients who had undergone surgery for breast cancer
}
\details{
\emph{Task:} formula(Survival~.)

\emph{From UCI:} "The dataset contains cases from a study that was conducted between 1958 and 1970 at the University of Chicago's Billings Hospital on the survival of patients who had undergone surgery for breast cancer."
}
\references{
Haberman, S. J. (1976). Generalized Residuals for Log-Linear Models, Proceedings of the 9th International Biometrics Conference, Boston, pp. 104-122.
}
\seealso{
\url{https://archive.ics.uci.edu/ml/datasets/Haberman's+Survival}
}

