% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createIndianLiverPatient}
\alias{createIndianLiverPatient}
\title{Create Indian Liver Patient dataset}
\usage{
createIndianLiverPatient(file = getfilepath("indianliverpatient.rds"),
  write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write data file to}

\item{write}{logical; should the dataset be written to disk for later use? (default: TRUE)}

\item{read}{logical; should we try to read the dataset from the specified location first? (default: TRUE)}
}
\value{
The dataset as a \code{\link[data.table]{data.table}}
}
\description{
This data set contains 416 liver patient records and 167 non liver patient records.The data set was collected from north east of Andhra Pradesh, India. Selector is a class label used to divide into groups(liver patient or not). This data set contains 441 male patient records and 142 female patient records. Any patient whose age exceeded 89 is listed as being of age "90".
}
\details{
Task:  formula(Patient ~ . )

1. Age  Age of the patient
2. Gender	Gender of the patient
3. TB	Total Bilirubin
4. DB	Direct Bilirubin
5. Alkphos Alkaline Phosphotase
6. Sgpt Alamine Aminotransferase
7. Sgot Aspartate Aminotransferase
8. TP	Total Protiens
9. ALB	Albumin
10. A/G Ratio	Albumin and Globulin Ratio
11. Selector field used to split the data into two sets (labeled by the experts)
}
\references{
Bendi Venkata Ramana, Prof. M. S. Prasad Babu and Prof. N. B. Venkateswarlu, Critical Comparative Study of Liver Patients from USA and INDIA: An Exploratory Analysis, International Journal of Computer Science Issues, ISSN :1694-0784, May 2012.
}
\seealso{
\url{https://archive.ics.uci.edu/ml/datasets/ILPD+(Indian+Liver+Patient+Dataset)}
}

