% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createMNIST}
\alias{createMNIST}
\title{Create MNIST digit recognition dataset}
\usage{
createMNIST(file = getfilepath("MNIST.rds"), write = TRUE, read = TRUE)
}
\description{
Code heavily inspired by https://gist.github.com/brendano/39760
}
\details{
First 60000 is train set, last 10000 is test set.

formula(Class~.)
}

