% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createParkinsons.R
\name{createParkinsons}
\alias{createParkinsons}
\title{Create Parkinsons dataset}
\usage{
createParkinsons(file = getfilepath("parkinsons.rds"), write = TRUE,
  read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write data file to}

\item{write}{logical; should the dataset be written to disk for later use? (default: TRUE)}

\item{read}{logical; should we try to read the dataset from the specified location first? (default: TRUE)}
}
\description{
TODO: dataset contains 32 subjects, UCI claims 31
}
\details{
Note there is a grouping in this dataset!
\emph{Task:} formula(status~ . -subject -recording)
}
\references{
Max A. Little, Patrick E. McSharry, Eric J. Hunter, Lorraine O. Ramig (2008), 'Suitability of dysphonia measurements for telemonitoring of Parkinson's disease', IEEE Transactions on Biomedical Engineering (to appear).
}

