% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSECOM.R
\name{createSECOM}
\alias{createSECOM}
\title{Create SECOM dataset}
\usage{
createSECOM(file = getfilepath("SECOM.rds"), write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write data file to}

\item{write}{logical; should the dataset be written to disk for later use? (default: TRUE)}

\item{read}{logical; should we try to read the dataset from the specified location first? (default: TRUE)}
}
\value{
The dataset as a \code{\link[data.table]{data.table}}
}
\description{
From UCI: " A complex modern semi-conductor manufacturing process is normally under consistent surveillance via the monitoring of signals/variables collected from sensors and or process measurement points. However, not all of these signals are equally valuable in a specific monitoring system. The measured signals contain a combination of useful information, irrelevant information as well as noise. It is often the case that useful information is buried in the latter two. Engineers typically have a much larger number of signals than are actually required. If we consider each type of signal as a feature, then feature selection may be applied to identify the most relevant signals. The Process Engineers may then use these signals to determine key factors contributing to yield excursions downstream in the process. This will enable an increase in process throughput, decreased time to learning and reduce the per unit production costs.
}
\details{
Key facts: Data Structure: The data consists of 2 files the dataset file SECOM consisting of 1567 examples each with 591 features a 1567 x 591 matrix and a labels file containing the classifications and date time stamp for each example. As with any real life data situations this data contains null values varying in intensity depending on the individuals features. This needs to be taken into consideration when investigating the data either through pre-processing or within the technique applied."

End goal is feature selection. Note that there might be constant features in the dataset. Task:  formula(Class ~ .)
}
\seealso{
\url{https://archive.ics.uci.edu/ml/datasets/SECOM}
}

