% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createSecStr}
\alias{createSecStr}
\title{Create SecStr dataset from the SSL benchmarks}
\usage{
createSecStr(file = getfilepath("secstr.rds"), write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write RData file to}

\item{write}{logical; should the dataset be written to disk? (default: FALSE)}
}
\description{
Task:
 formula(Class ~ .)
}
\details{
Cite: Chapelle
}
\seealso{
\url{http://olivier.chapelle.cc/ssl-book/benchmarks.html}
}

