% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createSonar}
\alias{createSonar}
\title{Create Sonar dataset}
\usage{
createSonar(file = getfilepath("sonar.rds"), write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write RData file to}

\item{write}{logical; should the dataset be written to disk? (default: FALSE)}
}
\description{
\emph{UCI description}: " Each pattern is a set of 60 numbers in the range 0.0 to 1.0. Each number represents the energy within a particular frequency band, integrated over a certain period of time. The integration aperture for higher frequencies occur later in time, since these frequencies are transmitted later during the chirp. The label associated with each record contains the letter "R" if the object is a rock and "M" if it is a mine (metal cylinder)."
}
\details{
\emph{Task:} formula(Label ~ .)
}

