% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSpambase.R
\name{createSpambase}
\alias{createSpambase}
\title{Create Spambase dataset}
\usage{
createSpambase(file = getfilepath("spambase.rds"), write = TRUE,
  read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write data file to}

\item{write}{logical; should the dataset be written to disk for later use? (default: TRUE)}

\item{read}{logical; should we try to read the dataset from the specified location first? (default: TRUE)}
}
\value{
The dataset as a \code{\link[data.table]{data.table}}
}
\description{
Task: formula(Class ~ .)
}
\seealso{
\url{https://archive.ics.uci.edu/ml/datasets/Spambase}
}

