% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createSPECT}
\alias{createSPECT}
\title{Create SPECT dataset}
\usage{
createSPECT(file = getfilepath("SPECT.rds"), write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write RData file to}

\item{write}{logical; should the dataset be written to disk? (default: FALSE)}
}
\description{
Single-photon emission computed tomography
}
\details{
First 80 train, the rest is test

formula(Diagnosis ~ .)
}

