% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createSPECTF}
\alias{createSPECTF}
\title{Create SPECTF dataset}
\usage{
createSPECTF(file = "SPECTF.rds", write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write RData file to}

\item{write}{logical; should the dataset be written to disk? (default: FALSE)}
}
\description{
Single-photon emission computed tomography
}
\details{
First 80 train, the rest is test

formula(Diagnosis ~ .)
}

