% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createSSLText}
\alias{createSSLText}
\title{Create SSL Text dataset from the SSL benchmarks}
\usage{
createSSLText(file = "SSLText.rds", write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write RData file to}

\item{write}{logical; should the dataset be written to disk? (default: FALSE)}
}
\description{
Task:
 formula(Class ~ .)
}
\details{
Cite: Chapelle
}

