% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTitanic.R
\name{createTitanic}
\alias{createTitanic}
\title{Create Titanic survival dataset}
\usage{
createTitanic(file = getfilepath("titanic.rds"), write = TRUE,
  read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write data file to}

\item{write}{logical; should the dataset be written to disk for later use? (default: TRUE)}

\item{read}{logical; should we try to read the dataset from the specified location first? (default: TRUE)}
}
\value{
The dataset as a \code{\link[data.table]{data.table}}
}
\description{
From source: "The titanic and titanic2 data frames describe the survival status of individual passengers on the Titanic. The titanic data frame does not contain information from the crew, but it does contain actual ages of half of the passengers. The principal source for data about Titanic passengers is the Encyclopedia Titanica. The datasets used here were begun by a variety of researchers. One of the original sources is Eaton & Haas (1994) Titanic: Triumph and Tragedy, Patrick Stephens Ltd, which includes a passenger list created by many researchers and edited by Michael A. Findlay."
}
\details{
Task:  formula(survived ~ . -name) # This includes life-boat information
}
\seealso{
\url{http://lib.stat.cmu.edu/S/Harrell/data/descriptions/titanic.html}
}

