% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWDBC.R
\name{createWDBC}
\alias{createWDBC}
\title{Create Wisconsin Breast Cancer Diagnosis}
\usage{
createWDBC(file = getfilepath("WDBC.rds"), write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write data file to}

\item{write}{logical; should the dataset be written to disk for later use? (default: TRUE)}

\item{read}{logical; should we try to read the dataset from the specified location first? (default: TRUE)}
}
\description{
From UCI:
Features are computed from a digitized image of a fine needle aspirate (FNA) of a breast mass.  They describe characteristics of the cell nuclei present in the image. A few of the images can be found at http://www.cs.wisc.edu/~street/images/
}
\details{
formula(Diagnosis ~ . -ID)
}

