% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createWPBC}
\alias{createWPBC}
\title{Create Wisconsin Breast Cancer Prognosis}
\usage{
createWPBC(file = "WPBC.RData", write = FALSE)
}
\arguments{
\item{file}{character; path/filename to write RData file to}

\item{write}{logical; should the dataset be written to disk? (default: FALSE)}
}
\description{
From UCI:
Features are computed from a digitized image of a fine needle aspirate (FNA) of a breast mass.  They describe characteristics of the cell nuclei present in the image. A few of the images can be found at http://www.cs.wisc.edu/~street/images/
}
\details{
Two tasks:
Predicting recurrence in next 24 months
formula(Outcome ~ . -ID -Time)

Predicting time of recurrence
Something like:
formula(Time ~ . -ID -Outcome)
}

