% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCrescentMoon.R
\name{generateCrescentMoon}
\alias{generateCrescentMoon}
\title{Generate Crescent Moon dataset}
\usage{
generateCrescentMoon(n = 100, sigma = 1)
}
\arguments{
\item{n}{Number of objects per class}

\item{sigma}{variance of the class}
}
\description{
Generate Crescent Moon dataset
}
\examples{
data<-generateCrescentMoon(150,1)
plot(data[,1],data[,2],col=data$Class,asp=1)
}

