% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generateGaussianClass}
\alias{generateGaussianClass}
\title{Create Classification problem with gaussian classes}
\usage{
generateGaussianClass(n = 100, d = 2, var = 1)
}
\description{
Task: formula(Class~.)
}
\examples{
d<-createGaussianClass(n=100,var=0.1)
plot(d$V1,d$V2,col=d$Class,asp=1)
}

