% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadir.R
\name{getfilepath}
\alias{getfilepath}
\title{Get full path by extending filename with the global data directory}
\usage{
getfilepath(filename)
}
\arguments{
\item{filename}{Filename in the data directory}
}
\description{
Get full path by extending filename with the global data directory
}

