% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDigit1.R
\name{createDigit1}
\alias{createDigit1}
\title{Create Digit1 dataset from the SSL benchmarks}
\usage{
createDigit1(file = getfilepath("digit1.rds"), write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write data file to}

\item{write}{logical; should the dataset be written to disk for later use? (default: TRUE)}

\item{read}{logical; should we try to read the dataset from the specified location first? (default: TRUE)}
}
\description{
Task:
 formula(Class ~ .)
}
\details{
Cite: Chapelle
}

