% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createG241N}
\alias{createG241N}
\title{Create G241C dataset from the SSL benchmarks}
\usage{
createG241N(file = getfilepath("g241n.rds"), write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write RData file to}

\item{write}{logical; should the dataset be written to disk? (default: FALSE)}
}
\description{
Task:
 formula(Class ~ .)
}
\details{
Cite: Chapelle
}

