% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMNIST.R
\name{createMNIST}
\alias{createMNIST}
\title{Create MNIST digit recognition dataset}
\usage{
createMNIST(file = getfilepath("MNIST.rds"), write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write data file to}

\item{write}{logical; should the dataset be written to disk for later use? (default: TRUE)}

\item{read}{logical; should we try to read the dataset from the specified location first? (default: TRUE)}
}
\description{
Code heavily inspired by https://gist.github.com/brendano/39760
}
\details{
First 60000 is train set, last 10000 is test set.

formula(Class~.)
}
\examples{
#as.matrix(data[1040,-785])  \%>\% matrix(28,28)  \%>\%  Matrix    \%>\% t\%>\% image(asp=1)
}

