% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createParkinsons}
\alias{createParkinsons}
\title{Create Parkinsons dataset}
\usage{
createParkinsons(file = "parkinsons.rds", write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write RData file to}

\item{write}{logical; should the dataset be written to disk? (default: FALSE)}
}
\description{

}
\details{
TODO: dataset contains 32 subjects, UCI claims 31

Note there is a grouping in this dataset!
\emph{Task:} formula(status~ . -subject -recording)
}
\references{
Max A. Little, Patrick E. McSharry, Eric J. Hunter, Lorraine O. Ramig (2008), 'Suitability of dysphonia measurements for telemonitoring of Parkinson's disease', IEEE Transactions on Biomedical Engineering (to appear).
}

