% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSPECTF.R
\name{createSPECTF}
\alias{createSPECTF}
\title{Create SPECTF dataset}
\usage{
createSPECTF(file = getfilepath("SPECTF.rds"), write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write data file to}

\item{write}{logical; should the dataset be written to disk for later use? (default: TRUE)}

\item{read}{logical; should we try to read the dataset from the specified location first? (default: TRUE)}
}
\description{
Single-photon emission computed tomography
}
\details{
First 80 train, the rest is test

formula(Diagnosis ~ .)
}

