% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createUSPS}
\alias{createUSPS}
\title{Create USPS dataset from the SSL benchmarks}
\usage{
createUSPS(file = "USPS.rds", write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write RData file to}

\item{write}{logical; should the dataset be written to disk? (default: FALSE)}
}
\description{
Task:
 formula(Class ~ .)
}
\details{
Cite: Chapelle
}

