% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createWDBC}
\alias{createWDBC}
\title{Create Wisconsin Breast Cancer Diagnosis}
\usage{
createWDBC(file = "WDBC.rds", write = TRUE, read = TRUE)
}
\arguments{
\item{file}{character; path/filename to write RData file to}

\item{write}{logical; should the dataset be written to disk? (default: FALSE)}
}
\description{
From UCI:
Features are computed from a digitized image of a fine needle aspirate (FNA) of a breast mass.  They describe characteristics of the cell nuclei present in the image. A few of the images can be found at http://www.cs.wisc.edu/~street/images/
}
\details{
formula(Diagnosis ~ . -ID)
}

