% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createWOBC}
\alias{createWOBC}
\title{Create Wisconsin Breast Cancer Diagnosis (Original)}
\usage{
createWOBC(file = getfilepath("WOBC.RData"), write = FALSE)
}
\arguments{
\item{file}{character; path/filename to write RData file to}

\item{write}{logical; should the dataset be written to disk? (default: FALSE)}
}
\description{
From UCI:
Citation Request:
formula(Class ~ . -ID)
}

