% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generateCrescentMoon}
\alias{generateCrescentMoon}
\title{Generate Crescent Moon dataset}
\usage{
generateCrescentMoon(n = 100, d = 2, var = 1)
}
\description{
Generate Crescent Moon dataset
}
\examples{
data<-generateCrescentMoon(150,1)
plot(data[,1],data[,2],col=data$Class,asp=1)
}

