% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateGaussianClass.R
\name{generateGaussianClass}
\alias{generateGaussianClass}
\title{Create Classification problem with gaussian classes}
\usage{
generateGaussianClass(n = 100, d = 2, var = 1)
}
\arguments{
\item{n}{Number of objects}

\item{d}{Number of features}

\item{var}{variance of the class}
}
\description{
Task: formula(Class~.)
}
\examples{
df <- generateGaussianClass(n=100,d=2,var=0.1)
plot(df$V1,df$V2,col=df$Class,asp=1)
}

